Flytte arkene til egne komponenter
