+++ src/komponenter/htmlPDF.tsx
@@ -1,7 +1,5 @@
 import React, { useContext } from "react";
-import { NavIkon } from "../ikoner/navikon";
 import "../stiler/htmlPDF.css";
-import parse from "html-react-parser";
 import { tabellObjekter } from "../mockdata/tabeller";
 import { SkjemaContext } from "../context/context";
 import {
@@ -10,7 +8,8 @@ import {
 	MAKS_ANTALL_LINJER,
 	settInnTabell,
 } from "../utils/htmlPdfUtils";
-import { avsnittType, skalAvsnittInkluderesType } from "../typer/typer";
+import { Ark } from "./ark";
+import { ArkMedBrevhode } from "./arkMedBrevhode";
 
 const HTMLPDF = () => {
 	const { avsnittState, brevmalTittelState, skalAvsnittInkluderesState } =
@@ -76,40 +75,6 @@ const HTMLPDF = () => {
 					<Ark innhold={innhold} />
 				)
 			)}
-		</div>
-	);
-};
-
-interface ArkProps {
-	innhold: string;
-}
-
-const Ark = ({ innhold }: ArkProps) => {
-	return (
-		<div className="ark">
-			<div className="innhold">
-				<>{parse(innhold)}</>
-			</div>
-		</div>
-	);
-};
-
-interface ArkMedBrevhodeProps {
-	brevmaltittel: string;
-	innhold: string;
-}
-
-const ArkMedBrevhode = ({ brevmaltittel, innhold }: ArkMedBrevhodeProps) => {
-	return (
-		<div className="ark">
-			<div className="brevhode">
-				<NavIkon />
-				<p>{new Date().toISOString().slice(0, 10)}</p>
-			</div>
-			<div className="innhold">
-				<h1 className="brevmaloverskrift">{brevmaltittel}</h1>
-				<>{parse(innhold)}</>
-			</div>
 		</div>
 	);
 };
